Improved build system
libVFRendering now builds on Debian 8 (Jessie) and Ubuntu 16.04 (Xenial)
using make and/or CMake. For Debian 8, a manual installation of a recent
CMake version is required, as Debian 8 only has CMake 3.0.2. Support for
older CMake versions will be looked into in the future.

The demo has been removed from the default targets of the Makefile
and for CMake, the BUILD_DEMO option was introduced. Currently the CI
systems do not build the demo.