Improved calculation of isosurface normals
Previously, only one "inside point" was used to determine the
direction of an normal vector. When two such points are available,
both should be used, as one of the two might be very close or even
directly on top of the triangle surface.

This commit should fix the issue of falsely flipped normals.